+++ src/Program.cs
@@ -73,7 +73,7 @@ internal class Program
             new AzureMcp.Areas.Sql.SqlSetup(),
             new AzureMcp.Areas.Storage.StorageSetup(),
             new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
-            new AzureMcp.Areas.TerraformBestPracticesForAzure.TerraformBestPracticesForAzureSetup(),
+            new AzureMcp.Areas.AzureTerraformBestPractices.AzureTerraformBestPracticesSetup(),
         ];
     }